+++ packages/sonarwhal/src/lib/utils/caniuse.ts
@@ -1,0 +1,3 @@
+import { isSupported } from 'caniuse-api';
+
+export { isSupported };