+++ packages/rule-disown-opener/CHANGELOG.md
@@ -1,3 +1,17 @@
+# 3.0.5 (June 6, 2018)
+
+## Bug fixes / Improvements
+
+* [[`4e5929c8e4`](https://github.com/sonarwhal/sonarwhal/commit/4e5929c8e4faab6431b63fe86045d4460fbaaf6d)] - Fix: Remove `pluralize` dependency (by [`Cătălin Mariș`](https://github.com/alrra) / see also: [`#1122`](https://github.com/sonarwhal/sonarwhal/issues/1122)).
+
+
+# 3.0.4 (May 31, 2018)
+
+## Bug fixes / Improvements
+
+* [[`ad0ba0223e`](https://github.com/sonarwhal/sonarwhal/commit/ad0ba0223ec03cab3a8b7f7bc69c8d63f72d7123)] - Fix: `BrowserslistError: Unknown version` issue (by [`Jesus David García Gomez`](https://github.com/sarvaje) / see also: [`#1109`](https://github.com/sonarwhal/sonarwhal/issues/1109)).
+
+
 # 3.0.3 (April 10, 2018)
 
 ## Bug fixes / Improvements